+++ tests/integration/test_merchant_account.py
@@ -37,94 +37,42 @@ class TestMerchantAccount(unittest.TestCase):
     }
 
     def test_return_all_merchant_accounts(self):
-        gateway = BraintreeGateway(
-            client_id="client_id$development$integration_client_id",
-            client_secret="client_secret$development$integration_client_secret"
-        )
+        result = TestHelper.get_merchant()
 
-        code = TestHelper.create_grant(gateway, {
-            "merchant_public_id": "integration_merchant_id",
-            "scope": "read_write"
-        })
-
-        result = gateway.oauth.create_token_from_code({
-            "code": code
-        })
-
-        gateway = BraintreeGateway(
-            access_token=result.credentials.access_token,
-            environment=Environment.Development
-        )
-
-        result = gateway.merchant_account.all()
-        merchant_accounts = [ma for ma in result.merchant_accounts.items]
-        self.assertTrue(len(merchant_accounts) > 20)
+        merchant_accounts = [ma for ma in result.merchant.merchant_accounts.items]
+        self.assertTrue(len(merchant_accounts) > 0)
 
     def test_returns_merchant_account_with_correct_attributes(self):
-        gateway = BraintreeGateway(
-            client_id="client_id$development$integration_client_id",
-            client_secret="client_secret$development$integration_client_secret"
-        )
+        result = TestHelper.get_merchant()
 
-        result = gateway.merchant.create({
-            "email": "name@email.com",
-            "country_code_alpha3": "GBR",
-            "payment_methods": ["credit_card", "paypal"]
-        })
-
-        gateway = BraintreeGateway(
-            access_token=result.credentials.access_token
-        )
-
-        result = gateway.merchant_account.all()
-        merchant_accounts = [ma for ma in result.merchant_accounts.items]
-        self.assertEqual(len(merchant_accounts), 1)
+        print(result.merchant.merchant_accounts)
+        merchant_accounts = [ma for ma in result.merchant.merchant_accounts.items]
+        self.assertTrue(len(merchant_accounts) > 0)
 
         merchant_account = merchant_accounts[0]
-        self.assertEqual(merchant_account.currency_iso_code, "GBP")
         self.assertEqual(merchant_account.status, MerchantAccount.Status.Active)
-        self.assertTrue(merchant_account.default)
 
     def test_find_404(self):
         with self.assertRaises(NotFoundError):
             MerchantAccount.find("not_a_real_id")
-
+    
     def test_merchant_account_create_for_currency(self):
-        self.gateway = BraintreeGateway(
-            client_id="client_id$development$integration_client_id",
-            client_secret="client_secret$development$integration_client_secret"
-        )
-
-        result = self.gateway.merchant.create({
-            "email": "name@email.com",
-            "country_code_alpha3": "GBR",
-            "payment_methods": ["credit_card", "paypal"]
-        })
+        result = TestHelper.get_merchant()
 
         gateway = BraintreeGateway(
             access_token=result.credentials.access_token,
         )
 
+        merchant_accounts = [ma for ma in result.merchant.merchant_accounts.items]
         result = gateway.merchant_account.create_for_currency({
-            "currency": "USD",
-            "id": "custom_id"
+            "currency": "JPY"
         })
 
         self.assertTrue(result.is_success)
-        self.assertEqual(result.merchant_account.currency_iso_code, "USD")
-        self.assertEqual(result.merchant_account.id, "custom_id")
+        self.assertEqual(result.merchant_account.currency_iso_code, "JPY")
 
     def test_merchant_account_create_for_currency_handles_invalid_currency(self):
-        self.gateway = BraintreeGateway(
-            client_id="client_id$development$integration_client_id",
-            client_secret="client_secret$development$integration_client_secret"
-        )
-
-        result = self.gateway.merchant.create({
-            "email": "name@email.com",
-            "country_code_alpha3": "GBR",
-            "payment_methods": ["credit_card", "paypal"]
-        })
+        result = TestHelper.get_merchant()
 
         gateway = BraintreeGateway(
             access_token=result.credentials.access_token,
@@ -138,16 +86,7 @@ class TestMerchantAccount(unittest.TestCase):
         self.assertEqual(result.errors.for_object("merchant").on("currency")[0].code, ErrorCodes.Merchant.CurrencyIsInvalid)
 
     def test_merchant_account_create_for_currency_handles_currency_requirement(self):
-        self.gateway = BraintreeGateway(
-            client_id="client_id$development$integration_client_id",
-            client_secret="client_secret$development$integration_client_secret"
-        )
-
-        result = self.gateway.merchant.create({
-            "email": "name@email.com",
-            "country_code_alpha3": "GBR",
-            "payment_methods": ["credit_card", "paypal"]
-        })
+        result = TestHelper.get_merchant()
 
         gateway = BraintreeGateway(
             access_token=result.credentials.access_token,
@@ -163,47 +102,37 @@ class TestMerchantAccount(unittest.TestCase):
         self.assertEqual(merchant_account.currency_iso_code, "USD")
         
     def test_merchant_account_create_for_currency_merchant_account_already_existing_for_currency(self):
-        self.gateway = BraintreeGateway(
-            client_id="client_id$development$integration_client_id",
-            client_secret="client_secret$development$integration_client_secret"
-        )
-
-        result = self.gateway.merchant.create({
-            "email": "name@email.com",
-            "country_code_alpha3": "GBR",
-            "payment_methods": ["credit_card", "paypal"]
-        })
+        result = TestHelper.get_merchant()
 
         gateway = BraintreeGateway(
             access_token=result.credentials.access_token,
         )
 
         result = gateway.merchant_account.create_for_currency({
-            "currency": "GBP",
+            "currency": "EUR",
+        })
+        self.assertTrue(result.is_success)
+
+        result = gateway.merchant_account.create_for_currency({
+            "currency": "EUR",
         })
 
         self.assertFalse(result.is_success)
         self.assertEqual(result.errors.for_object("merchant").on("currency")[0].code, ErrorCodes.Merchant.MerchantAccountExistsForCurrency)
-
+    
+  
     def test_merchant_account_create_for_currency_merchant_account_already_existing_for_id(self):
-        self.gateway = BraintreeGateway(
-            client_id="client_id$development$integration_client_id",
-            client_secret="client_secret$development$integration_client_secret"
-        )
-
-        result = self.gateway.merchant.create({
-            "email": "name@email.com",
-            "country_code_alpha3": "GBR",
-            "payment_methods": ["credit_card", "paypal"]
-        })
+        result = TestHelper.get_merchant()
 
         gateway = BraintreeGateway(
             access_token=result.credentials.access_token,
         )
 
+        merchant_accounts = [ma for ma in result.merchant.merchant_accounts.items]
+
         result = gateway.merchant_account.create_for_currency({
             "currency": "GBP",
-            "id": result.merchant.merchant_accounts[0].id
+            "id": merchant_accounts[0].id
         })
 
         self.assertFalse(result.is_success)